# openapi.yaml
openapi: 3.0.2
info:
  version: 1.0.0
  title: payments-configuration-management-service OpenAPI Documentation
  x-gateway-info:
    name: payments-configuration-management-service
...
paths:
...
  /api/countries:
    get:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - Country
      parameters:
        - in: query
          name: authority
          required: false
          schema:
            $ref: "#/components/schemas/Authority"
        - in: query
          name: providerCode
          required: false
          schema:
            type: string
      summary: "Get country configs"
      description: "Returns an array of country configs for countries from all or specified authority"
      operationId: "getCountries"
      x-eov-operation-handler: "api/methods"
      responses:
        200:
          description: "Successful Request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  countries:
                    type: array
                    items:
                      $ref: "#/components/schemas/CountryConfig"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"

    put:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - Country
      summary: "Add a new country"
      description: "Add a new country to the database"
      operationId: "addCountry"
      x-eov-operation-handler: "api/methods"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ['iso2', 'iso3', 'name']
              additionalProperties: false
              properties:
                iso2:
                  type: string
                  minLength: 2
                  maxLength: 2
                  pattern: '^[A-Z]{2}$'
                iso3:
                  type: string
                  minLength: 3
                  maxLength: 3
                  pattern: '^[A-Z]{3}$'
                name:
                  type: string
                  minLength: 1
                  maxLength: 60
                group:
                  type: string
                  maxLength: 30
      responses:
        200:
          description: Successful Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        422:
          $ref: "#/components/responses/ValidationError"
        500:
          $ref: "#/components/responses/ServerError"
...
  /api/providers/{code}/methods/filter:
    post:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - Provider
      summary: "Get payment methods with state by provider"
      description: "Returns methods with state by provider code"
      operationId: "getProviderBoundedMethods"
      x-eov-operation-handler: "api/methods"
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                countriesAuthorities:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/CountryAuthorityDto'
      responses:
        200:
          description: Successful Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderMethodBoundedDto"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
...
  /api/providers/{code}/restrictions:
    get:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - Provider
      summary: "Get provider platform versions"
      operationId: "getProviderRestrictions"
      x-eov-operation-handler: "api/methods"
      description: "Returns provider platform versions restrictions"
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Successful Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderRestrictions'
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
    put:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - Provider
      summary: "Update provider restrictions"
      operationId: "updateProviderRestrictions"
      x-eov-operation-handler: "api/methods"
      description: "Update provider restrictions for mobile platforms"
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderRestrictions'
      responses:
        200:
          description: 'Successful Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderRestrictions'
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        409:
          $ref: "#/components/responses/CredentialsOverlapError"
  /api/provider-methods/withdrawals-order:
    get:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - Provider Method
      summary: "Get Provider Methods sorted by order"
      operationId: "getWithdrawalsOrder"
      x-eov-operation-handler: "api/methods"
      description: |
        The endpoint returns enabled ProviderMethods for payouts and refunds in a selected authority and country in the DESC order according to the payoutOrder and refundOrder
      parameters:
        - in: query
          name: authority
          required: true
          schema:
            $ref: "#/components/schemas/Authority"
        - in: query
          name: country
          required: true
          schema:
            $ref: "#/components/schemas/CountryIso2"
      responses:
        200:
          description: 'Successful Request'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalsOrder"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
    put:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - Provider Method
      summary: "Update Provider Methods withdrawals order"
      operationId: "updateWithdrawalsOrder"
      x-eov-operation-handler: "api/methods"
      description: |
        The endpoint updates refunds and payout order values and enabled ProviderMethods for payouts and refunds in a selected authority and country in the DESC order after update
      parameters:
        - in: query
          name: authority
          required: true
          schema:
            $ref: "#/components/schemas/Authority"
        - in: query
          name: country
          required: true
          schema:
            $ref: "#/components/schemas/CountryIso2"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalsOrder"
      responses:
        200:
          description: 'Successful Request'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalsOrder"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        409:
          $ref: "#/components/responses/Conflict"
        500:
          $ref: "#/components/responses/ServerError"
  /api/country-authority:
    put:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - CountryAuthority
      summary: "Create country authority"
      description: "This endpoint creates a new country authority. Once created, it returns the details of the new country authority."
      operationId: "createCountryAuthority"
      x-eov-operation-handler: "api/methods"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CountryAuthorityDto"
      responses:
        200:
          description: Successful Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryAuthorityDto"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        409:
          $ref: "#/components/responses/Conflict"
        500:
          $ref: "#/components/responses/ServerError"
  /api/country-authority-methods:
    get:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - CountryAuthorityMethod
      parameters:
        - in: query
          name: authority
          required: true
          schema:
            $ref: "#/components/schemas/Authority"
        - in: query
          name: country
          required: true
          schema:
            $ref: "#/components/schemas/CountryIso2"
      summary: "Get payment method configs"
      description: "Returns an array of CountryAuthorityMethods with provider names by authority name and country ISO-2"
      operationId: "getCountryAuthorityMethods"
      x-eov-operation-handler: "api/methods"
      responses:
        200:
          description: Successful Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentMethodConfigs:
                    type: array
                    items:
                      $ref: "#/components/schemas/PaymentMethodConfig"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
  /api/country-authority-methods/{methodCode}/status:
    put:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - CountryAuthorityMethod
      parameters:
        - in: query
          name: authority
          required: true
          schema:
            $ref: "#/components/schemas/Authority"
        - in: query
          name: country
          required: true
          schema:
            $ref: "#/components/schemas/CountryIso2"
        - in: path
          name: methodCode
          required: true
          schema:
            $ref: "#/components/schemas/MethodCode"
      summary: "Update payment method status"
      description: "Updates payment method status value by its code for specified authority name and country ISO-2"
      operationId: "updateCountryAuthorityMethodStatus"
      x-eov-operation-handler: "api/methods"
      requestBody:
        $ref: "#/components/requestBodies/UpdateCountryAuthorityMethodStatus"
      responses:
        200:
          description: 'Returns successfully updated status value'
          content:
            application/json:
              schema:
                type: object
                properties:
                  isEnabled:
                    type: boolean
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
  /api/country-authority-methods/order:
    put:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - CountryAuthorityMethod
      parameters:
        - in: query
          name: country
          required: true
          schema:
            $ref: "#/components/schemas/CountryIso2"
        - in: query
          name: authority
          required: true
          schema:
            $ref: "#/components/schemas/Authority"
      summary: "Update order of CountryAuthorityMethods"
      description: "Update order of CountryAuthorityMethods for deposits"
      operationId: "updateCountryAuthorityMethodsOrder"
      x-eov-operation-handler: "api/methods"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: ['methodCodes']
              properties:
                methodCodes:
                  type: array
                  items:
                    type: string
                    minLength: 1
      responses:
        200:
          description: "Orders successfully updated"
          content:
            application/json:
              schema:
                type: object
                required: ['methodCodes']
                properties:
                  methodCodes:
                    type: array
                    items:
                      type: string
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        409:
          $ref: "#/components/responses/Conflict"
        500:
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
  examples:
    ERR_CREDENTIALS_OVERLAP:
      value:
        code: ERR_CREDENTIALS_OVERLAP
        message: "CredentialsOverlapError"
        requestId: "6db8a776-6dff-400d-8baf-3a6c81b45eff"
        meta:
          source: 0
          conflict: ['one']
    ERR_PARAMETERS_OVERLAP:
      value:
        code: ERR_PARAMETERS_OVERLAP
        message: "CredentialsOverlapError"
        requestId: "6db8a776-6dff-400d-8baf-3a6c81b45eff"
        meta:
          source: 0
          target: 1
          conflict: ['one', 'two']
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        meta:
          type: object
      required:
        - code
        - requestId
    Authority:
      type: string
      minLength: 1
      maxLength: 20
      example: 'FSCM'
...
    CountryIso2:
      type: string
      pattern: ^[A-Za-z]{2}$
      minLength: 2
      maxLength: 2
      example: "CY"
...
    MethodCode:
      minLength: 1
      maxLength: 50
      type: string
...
    CountryConfig:
      type: object
      properties:
        countryGroup:
          type: string
        countryData:
          type: object
          properties:
            countryCode:
              type: string
            countryName:
              type: string
            authorities:
              type: array
              items:
                $ref: "#/components/schemas/Authority"
    Country:
      type: object
      required: ['iso2', 'iso3', 'name']
      properties:
        iso2:
          type: string
          minLength: 2
          maxLength: 2
          pattern: '^[A-Z]{2}$'
        iso3:
          type: string
          minLength: 3
          maxLength: 3
          pattern: '^[A-Z]{3}$'
        name:
          type: string
          minLength: 1
          maxLength: 60
        group:
          type: string
          maxLength: 30
...
    CountryAuthorityDto:
      type: object
      required: [ 'country', 'authority' ]
      additionalProperties: false
      properties:
        country:
          $ref: '#/components/schemas/CountryIso2'
        authority:
          $ref: '#/components/schemas/Authority'
...
    WithdrawalsOrder:
      type: object
      required: ['refunds', 'payouts']
      additionalProperties: false
      properties:
        refunds:
          type: array
          items:
            $ref: '#/components/schemas/ProviderMethod'
        payouts:
          type: array
          items:
            $ref: '#/components/schemas/ProviderMethod'
    ProviderMethod:
      type: object
      required: ['providerCode', 'methodCode']
      additionalProperties: false
      properties:
        providerCode:
          type: string
          minLength: 1
          maxLength: 100
        methodCode:
          type: string
          minLength: 1
          maxLength: 50
        providerName:
          type: string
          minLength: 1
          maxLength: 100
        methodName:
          type: string
          minLength: 1
          maxLength: 100
...
    ProviderMethodBoundedDto:
      type: object
      properties:
        methodCode:
          type: string
        methodName:
          type: string
        state:
          enum:
            - bounded
            - mixed
            - not_bounded
    PaymentMethodConfig:
      type: object
      properties:
        methodCode:
          type: string
        methodName:
          type: string
        isEnabled:
          type: boolean
        providers:
          type: array
          description: Provider names
          items:
            type: string
...
    ApplicationPlatforms:
      type: string
      enum:
        - web
        - android
        - ios
    ProviderRestrictions:
      type: array
      items:
        type: object
        required: ['platform', 'isEnabled', 'countriesAuthorities', 'settings']
        properties:
          platform:
            $ref: '#/components/schemas/ApplicationPlatforms'
          isEnabled:
            type: boolean
          countriesAuthorities:
            type: array
            minItems: 1
            items:
              $ref: '#/components/schemas/CountryAuthorityDto'
          settings:
            type: array
            items:
              $ref: '#/components/schemas/RestrictionSetting'
    RestrictionSetting:
      type: object
      required: ['condition', 'version']
      properties:
        condition:
          type: string
          enum:
            - 'gte'
            - 'lte'
            - 'eq'
        version:
          type: string
          pattern: ^\d+\.\d+\.\d+$
...
  requestBodies:
    UpdateCountryAuthorityMethodStatus:
      content:
        application/json:
          schema:
            type: object
            required: ['isEnabled']
            additionalProperties: false
            properties:
              isEnabled:
                type: boolean
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    CredentialsOverlapError:
      description: 'CredentialsOverlapError'
      content:
        application/json:
          examples:
            ERR_CREDENTIALS_OVERLAP:
              $ref: '#/components/examples/ERR_CREDENTIALS_OVERLAP'
            ERR_PARAMETERS_OVERLAP:
              $ref: '#/components/examples/ERR_PARAMETERS_OVERLAP'
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServerError:
      description: Server error.